requirements.txt
absl-py==0.8.1
asgiref==3.2.3
astor==0.8.0
attrs==19.3.0
autopep8==1.4.4
backcall==0.1.0
bleach==3.1.0
cachetools==3.1.1
certifi==2019.11.28
chardet==3.0.4
decorator==4.4.1
defusedxml==0.6.0
dj-database-url==0.5.0
Django==3.0.2
django-crispy-forms==1.8.1
django-widget-tweaks==1.4.5
editdistance==0.5.3
entrypoints==0.3
flake8==3.7.9
gast==0.2.2
google-api-core==1.14.3
google-auth==1.7.1
google-auth-oauthlib==0.4.1
google-cloud-vision==0.41.0
google-pasta==0.1.8
googleapis-common-protos==1.6.0
grpcio==1.25.0
gunicorn==20.0.4
h5py==2.10.0
html5lib==1.0.1
idna==2.8
importlib-metadata==1.2.0
ipykernel==5.1.3
ipython==7.10.1
ipython-genutils==0.2.0
ipywidgets==7.5.1
jedi==0.15.1
Jinja2==2.10.3
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==5.3.4
jupyter-console==6.0.0
jupyter-core==4.6.1
kaldiio==2.15.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
llvmlite==0.30.0
Markdown==3.1.1
MarkupSafe==1.1.1
mccabe==0.6.1
mistune==0.8.4
more-itertools==8.0.0
nbconvert==5.6.1
nbformat==4.4.0
notebook==6.0.2
numba==0.46.0
numpy==1.17.4
oauthlib==3.1.0
opencv-python==4.1.2.30
opt-einsum==3.1.0
pandocfilters==1.4.2
parso==0.5.1
pexpect==4.7.0
pickleshare==0.7.5
Pillow==7.0.0
postgres==3.0.0
prometheus-client==0.7.1
prompt-toolkit==2.0.10
protobuf==3.11.1
psycopg2==2.8.4
psycopg2-binary==2.8.4
psycopg2-pool==1.1
ptyprocess==0.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.7
pycodestyle==2.5.0
pyflakes==2.1.1
Pygments==2.5.2
PyPDF2==1.26.0
pyrsistent==0.15.6
python-dateutil==2.8.1
pytz==2019.3
pyzmq==18.1.1
qtconsole==4.6.0
reportlab==3.5.34
requests==2.22.0
requests-oauthlib==1.3.0
rsa==4.0
Send2Trash==1.5.0
six==1.13.0
sqlparse==0.3.0
tensorboard==1.15.0
tensorflow-estimator==1.15.1
tensorflow-gpu==1.15.0
termcolor==1.1.0
terminado==0.8.3
testpath==0.4.4
tornado==6.0.3
traitlets==4.3.3
urllib3==1.25.7
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.16.0
widgetsnbextension==3.5.1
wrapt==1.11.2
xhtml2pdf==0.2.4
zipp==0.6.0
